refactor(prompt-template): split core module into prompt and string
