+++ src/Langchain/Chain/MapReduce.hs
@@ -37,7 +37,7 @@ import Langchain.Core.Model
   , userMessage
   )
 import Langchain.DocumentLoader.Core (Document (..))
-import Langchain.PromptTemplate (PromptTemplate, fromTemplate, renderPrompt)
+import Langchain.PromptTemplate.Prompt (PromptTemplate, fromTemplate, renderPrompt)
 
 -- | Map-Reduce chain configuration
 data MapReduceChain model = MapReduceChain
