+++ src/Langchain/Prelude.hs
@@ -477,8 +477,8 @@ import Langchain.Chain.Summarization
 import Langchain.Config.Validation
 import Langchain.Core.Error
 import Langchain.Core.Model
-import qualified Langchain.Core.Monad as CoreMonad
 import Langchain.Core.Monad
+import qualified Langchain.Core.Monad as CoreMonad
 import Langchain.Core.Runnable
 import Langchain.Core.Stream
 import Langchain.Core.Tool
@@ -522,7 +522,8 @@ import Langchain.OutputParser.Enum
 import Langchain.OutputParser.Structured
 import Langchain.OutputParser.Xml
 import Langchain.Pipeline.DSL
-import Langchain.PromptTemplate
+import Langchain.PromptTemplate.FewShot
+import Langchain.PromptTemplate.Prompt
 import Langchain.Provider.Anthropic (Anthropic, newAnthropic)
 import Langchain.Provider.DeepSeek (DeepSeek, newDeepSeek)
 import Langchain.Provider.Gemini (Gemini, newGemini)
@@ -544,10 +545,10 @@ import Langchain.TextSplitter.Markdown
 import Langchain.TextSplitter.RecursiveCharacter
 import Langchain.TextSplitter.Token
 import Langchain.Tool.Async
-import Langchain.VectorStore.Filter
 import Langchain.Tool.GenericSchema
 import Langchain.Trace.Core
 import Langchain.VectorStore.Core
+import Langchain.VectorStore.Filter
 import Langchain.VectorStore.InMemory
 import Langchain.VectorStore.PgVector
 import Langchain.VectorStore.Qdrant
