+++ src/Langchain/Retriever/ContextualCompression.hs
@@ -35,7 +35,7 @@ import Langchain.Core.Model
   , userMessage
   )
 import Langchain.DocumentLoader.Core (Document (..))
-import Langchain.PromptTemplate (PromptTemplate, fromTemplate, renderPrompt)
+import Langchain.PromptTemplate.Prompt (PromptTemplate, fromTemplate, renderPrompt)
 import Langchain.Retriever.Core (Retriever (..))
 
 -- | Contextual compression retriever container
